requirements.txt
aiohttp~=3.8.6
aiomysql~=0.2.0
cryptography~=41.0.5
discord~=2.3.2
orjson~=3.9.10
python-dotenv~=1.0.0
pyright~=1.1.333
black~=23.10.1
isort~=5.12.0
requests~=2.31.0